Move non-invite specific stuff to UserAddress
