+++ src/utils/MultiInviter.js
@@ -14,7 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {getAddressType, inviteToRoom} from '../Invite';
+import {getAddressType} from '../UserAddress';
+import {inviteToRoom} from '../Invite';
 import Promise from 'bluebird';
 
 /**
